+++ src/Request/GetB2BAccountFields.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Easir\SDK\Request;
+
+class GetB2BAccountFields extends GetFields
+{
+    /**
+     * @var string
+     */
+    protected $urlSuffix = '/account';
+}