+++ src/Request/GetB2CContactFields.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Easir\SDK\Request;
+
+class GetB2CContactFields extends GetFields
+{
+    /**
+     * @var string
+     */
+    protected $urlSuffix = '/b2c_contact';
+}